requirements.txt
face_recognition>=1.3.0
pillow>=9.0.0
numpy>=1.20.0
tqdm>=4.64.0
psutil>=5.9.0